pyproject.toml
[project]
name = "slack-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.27.0",
    "slack-sdk>=3.39.0",
    "uvicorn>=0.38.0",
]
